requirements.txt
fastapi
uvicorn[standard]
pydantic
transformers
torch
python-docx
PyMuPDF
